Moved ApplicationDbContext inside Adapters namespace
