+++ Registry.Web/Services/Adapters/ApplicationDbContext.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using IdentityServer4.EntityFramework.Options;
-using Microsoft.AspNetCore.ApiAuthorization.IdentityServer;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Options;
+using Registry.Web.Models;
 
-namespace Registry.Web.Models
+namespace Registry.Web.Services.Adapters
 {
     public sealed class ApplicationDbContext : IdentityDbContext<User>
     {
